+++ goldens/public-api/core/global_utils.api.md
@@ -4,6 +4,8 @@
 
 ```ts
 
+import { ɵFramework } from '@angular/core';
+
 // @public
 export function applyChanges(component: {}): void;
