+++ packages/core/BUILD.bazel
@@ -36,10 +36,10 @@ ng_project(
         "//:node_modules/zone.js",
         "//packages:goog_types",
         "//packages/core/primitives/defer:triggers",
+        "//packages/core/primitives/devtools",
         "//packages/core/primitives/di",
         "//packages/core/primitives/dom-navigation",
         "//packages/core/primitives/event-dispatch",
-        "//packages/core/primitives/profiler",
         "//packages/core/primitives/signals",
         "//packages/core/src/compiler",
         "//packages/core/src/di/interface",
