+++ packages/core/primitives/devtools/README.md
@@ -1,0 +1,3 @@
+# Angular Devtools
+
+This directory contains interfaces and enums related to the [Angular Devtools](https://angular.dev/tools/devtools). Use these interfaces and enums if you need to hook into Angular Devtools.
