+++ packages/core/primitives/devtools/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.dev/license
+ */
+
+export {ProfilerEvent, type Profiler} from './src/profiler_types';
+export {Framework} from './src/framework';
