--- packages/core/primitives/profiler/README.md
@@ -1,3 +1,0 @@
-# Angular Profiler Types
-
-This directory contains types related to the [Angular Devtools Profiler](https://angular.dev/tools/devtools/profiler). These types are for code that needs to hook into the Angular Devtools Profiler lifecycle.
