+++ packages/core/src/application/application_ref.ts
@@ -45,7 +45,7 @@ import {TESTABILITY} from '../testability/testability';
 import {NgZone} from '../zone/ng_zone';
 
 import {profiler} from '../render3/profiler';
-import {ProfilerEvent} from '../../primitives/profiler/src/profiler_types';
+import {ProfilerEvent} from '../../primitives/devtools';
 import {EffectScheduler} from '../render3/reactivity/root_effect_scheduler';
 import {isReactiveLViewConsumer} from '../render3/reactive_lview_consumer';
 import {ApplicationInitStatus} from './application_init';
