+++ packages/core/src/render3/util/discovery_utils.ts
@@ -26,6 +26,7 @@ import {CLEANUP, CONTEXT, LView, TVIEW, TViewType} from '../interfaces/view';
 
 import {getRootContext} from './view_traversal_utils';
 import {getLViewParent, unwrapRNode} from './view_utils';
+import {Framework} from '../../../primitives/devtools';
 
 /**
  * Retrieves the component instance associated with a given DOM element.
@@ -222,13 +223,6 @@ export function getDirectives(node: Node): {}[] {
   // The `directives` in this case are a named array called `LComponentView`. Clone the
   // result so we don't expose an internal data structure in the user's console.
   return context.directives === null ? [] : [...context.directives];
-}
-
-/** The framework used to author a particular application or component. */
-export enum Framework {
-  Angular = 'angular',
-  ACX = 'acx',
-  Wiz = 'wiz',
 }
 
 /** Metadata common to directives from all frameworks.  */
