+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -205,6 +205,7 @@
       "PendingTasksInternal",
       "PlatformRef",
       "PristineChangeEvent",
+      "ProfilerEvent",
       "QUERIES",
       "R3Injector",
       "R3ViewContainerRef",
