+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -201,6 +201,7 @@
       "PendingTasksInternal",
       "PlatformRef",
       "PristineChangeEvent",
+      "ProfilerEvent",
       "QUERIES",
       "R3Injector",
       "R3ViewContainerRef",
