+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -177,6 +177,7 @@
       "PRESERVE_HOST_CONTENT",
       "PRESERVE_HOST_CONTENT_DEFAULT",
       "PendingTasksInternal",
+      "ProfilerEvent",
       "QUERIES",
       "R3Injector",
       "REACTIVE_LVIEW_CONSUMER_NODE",
