+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -148,6 +148,7 @@
       "PRESERVE_HOST_CONTENT",
       "PRESERVE_HOST_CONTENT_DEFAULT",
       "PendingTasksInternal",
+      "ProfilerEvent",
       "QUERIES",
       "R3Injector",
       "REACTIVE_LVIEW_CONSUMER_NODE",
